feat: implemented gc::Type for RawTypeInfo
