+++ crates/mun_gc/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "mun_gc"
+version = "0.2.0"
+authors = ["The Mun Team <team@mun-lang.org>"]
+edition = "2018"
+homepage = "https://mun-lang.org"
+repository = "https://github.com/mun-lang/mun"
+license = "MIT OR Apache-2.0"
+description = "Garbage collection crate for Mun"
+
+[dependencies]
+abi = { path = "../mun_abi", package = "mun_abi" }
