+++ crates/mun_runtime/Cargo.toml
@@ -10,6 +10,7 @@ description = "A runtime for hot reloading and invoking Mun from Rust"
 
 [dependencies]
 abi = { path = "../mun_abi", package = "mun_abi" }
+gc = { path = "../mun_gc", package = "mun_gc" }
 failure = "0.1.5"
 libloading = "0.5"
 md5= "0.7.0"
